Makefile automated push

+++ libft/ft_power.c
@@ -6,13 +6,15 @@
 /*   By: ldevelle <ldevelle@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/01/07 14:30:26 by ldevelle          #+#    #+#             */
-/*   Updated: 2019/01/07 14:30:29 by ldevelle         ###   ########.fr       */
+/*   Updated: 2019/02/14 13:41:50 by ldevelle         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-int		ft_power(int nb, int power)
+#include "libft.h"
+
+intmax_t		ft_power(int nb, int power)
 {
-	int x;
+	intmax_t x;
 
 	x = 1;
 	if (power == 0)
